requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
defusedxml==0.7.1
Django==4.2.3
django-js-asset==2.1.0
django-mptt==0.14.0
django-tasks @ file:///D:/Projects/Python/Django/django-tasks/dist/django-tasks-0.1.tar.gz#sha256=956d6b971a7d8ab35bae62afdefd508dd341ca1af65719cbedfb479ea4692670
djangorestframework==3.14.0
drf-spectacular==0.26.4
EasyProcess==1.1
entrypoint2==1.1
httpie==3.2.2
idna==3.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
mdurl==0.1.2
MouseInfo==0.1.3
mss==9.0.1
multidict==6.0.4
mysqlclient==2.2.0
Pillow==10.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.16.1
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pytweening==1.0.7
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.2
rpds-py==0.10.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
